requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
appdirs==1.4.4
    # via virtualenv
attrs==20.3.0
    # via pytest
click==7.1.2
    # via code-annotations
code-annotations==1.1.0
    # via -r requirements/test.in
coverage==5.5
    # via pytest-cov
distlib==0.3.1
    # via virtualenv
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   code-annotations
filelock==3.0.12
    # via
    #   tox
    #   virtualenv
iniconfig==1.1.1
    # via pytest
jinja2==2.11.3
    # via code-annotations
markupsafe==1.1.1
    # via jinja2
packaging==20.9
    # via
    #   pytest
    #   tox
pbr==5.5.1
    # via stevedore
pluggy==0.13.1
    # via
    #   pytest
    #   tox
py==1.10.0
    # via
    #   pytest
    #   tox
pyparsing==2.4.7
    # via packaging
pytest-cov==2.11.1
    # via -r requirements/test.in
pytest-django==4.1.0
    # via -r requirements/test.in
pytest==6.2.2
    # via
    #   pytest-cov
    #   pytest-django
python-slugify==4.0.1
    # via code-annotations
pytz==2021.1
    # via
    #   -r requirements/base.txt
    #   django
pyyaml==5.4.1
    # via code-annotations
six==1.15.0
    # via
    #   tox
    #   virtualenv
sqlparse==0.4.1
    # via
    #   -r requirements/base.txt
    #   django
stevedore==3.3.0
    # via code-annotations
text-unidecode==1.3
    # via python-slugify
toml==0.10.2
    # via
    #   pytest
    #   tox
tox==3.23.0
    # via -r requirements/test.in
virtualenv==20.4.2
    # via tox
